+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Forum threads that need activation trigger a bell notification #289 !3703 @alex.simm
 - Improved map marker caching #2106 !3712 @AntonBallmaier
 - Added a map marker selection for ambassadors to display users #1759 !3714 @AntonBallmaier
+- Added support form for our new zammad ticket system !3210 @chriswalg @alex.simm
 
 ## Changes
 - Improve accessibility in create forum thread form !3488 @iron9